% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climextRemes-package.R
\docType{package}
\name{climextRemes-package}
\alias{climextRemes}
\alias{climextRemes-package}
\title{climextRemes: Tools for Analyzing Climate Extremes}
\description{
Functions for fitting GEV and POT (via point process fitting) models for extremes in climate data, providing return values, return probabilities, and return periods for stationary and nonstationary models. Also provides differences in return values and differences in log return probabilities for contrasts of covariate values. Functions for estimating risk ratios for event attribution analyses, including uncertainty. Under the hood, many of the functions use functions from 'extRemes', including for fitting the statistical models. Details are given in Paciorek, Stone, and Wehner (2018) \doi{10.1016/j.wace.2018.01.002}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/lbl-cascade/climextremes-dev}
}

}
\author{
\strong{Maintainer}: Christopher Paciorek \email{paciorek@stat.berkeley.edu}

Other contributors:
\itemize{
  \item Harinarayan Krishnan \email{hkrishnan@lbl.gov} [contributor]
}

}
\keyword{internal}
