% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{remove_runs}
\alias{remove_runs}
\title{Remove consecutive exceedances from a vector}
\usage{
remove_runs(y, index, upperTail = TRUE)
}
\arguments{
\item{y}{vector of values}

\item{index}{vector of indices, one per value, that indicate which elements of \code{y} are consecutive}

\item{upperTail}{logical indicating whether values of \code{y} are upper (right) tail values (TRUE) or lower (left) tail values (FALSE)}
}
\description{
Remove runs, i.e., consecutive exceedances, from a vector of values and associated indices (days); for use in declustering
}

