% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxil.R
\name{screen_within_block}
\alias{screen_within_block}
\title{Remove multiple exceedances within non-overlapping blocks of fixed length}
\usage{
screen_within_block(y, index, blockLength = 10)
}
\arguments{
\item{y}{vector of values}

\item{index}{vector of indices, one per value, that indicate which elements of \code{y} are consecutive}

\item{blockLength}{length of block within which to remove all but the most extreme value}
}
\description{
Remove multiple exceedances within non-overlapping blocks of fixed lengths, for use in declustering
}

