% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(vec, shift = NULL, lower = NULL, upper = NULL)
}
\arguments{
\item{vec}{vector of values}

\item{shift}{optional central point (if not provided, uses the mean of \code{vec})}

\item{lower}{optional lower end point of range (if not provided uses min of \code{vec})}

\item{upper}{optional upper end point of range (if not provided uses max of \code{vec})}
}
\description{
Normalize a vector by subtracting off central point and dividing by range
}

